function that reverses the content of an array of integers

+++ 0x06-pointers_arrays_strings/4-rev_array.c
@@ -5,7 +5,7 @@
  */
 void reverse_array(int *a, int n)
 {
-	short aux, i = 0;
+	int aux, i = 0;
 
 	for (; i < n; i++)
 	{
